refactered code

+++ frontend/src/v2/DeleteStatusModal.vue
@@ -1,34 +1,3 @@
-<template>
-  <div class="modal-mask" v-if="isOpen">
-    <div class="itbkk-message">
-      <h2 class="text-lg font-semibold mb-4">Delete Status</h2>
-      <p class="text-left mb-4">Are you sure you want to delete the status?</p>
-      <div class="flex justify-end">
-        <button
-          type="button"
-          class="px-4 py-2 bg-gray-300 text-gray-800 rounded-md mr-2 itbkk-button-cancel"
-          @click="closeModal"
-        >
-          Cancel
-        </button>
-        <button
-          type="button"
-          class="px-4 py-2 bg-red-500 text-white hover:bg-red-600 rounded-md itbkk-button-confirm"
-          @click="deleteStatus"
-        >
-          Confirm
-        </button>
-      </div>
-    </div>
-  </div>
-  <Toast
-    :show="showToast"
-    :statusCode="statusCode"
-    :operationType="operationType"
-    @close="showToast = false"
-  />
-</template>
-
 <script setup>
 import { defineProps, defineEmits, ref } from 'vue'
 import fetchUtils from '../lib/fetchUtils'
@@ -65,6 +34,12 @@ const deleteStatus = async () => {
       throw new Error('The "No Status" status cannot be deleted')
     }
 
+    if (props.statusIdToDelete === 6) {
+      alert('The "Done" status cannot be deleted')
+      closeModal()
+      throw new Error('The "No Status" status cannot be deleted')
+    }
+
     const response = await fetchUtils.deleteData(
       `statuses/${props.statusIdToDelete}`
     )
@@ -94,6 +69,37 @@ const deleteStatus = async () => {
   }
 }
 </script>
+
+<template>
+  <div class="modal-mask" v-if="isOpen">
+    <div class="itbkk-message">
+      <h2 class="text-lg font-semibold mb-4">Delete Status</h2>
+      <p class="text-left mb-4">Are you sure you want to delete the status?</p>
+      <div class="flex justify-end">
+        <button
+          type="button"
+          class="px-4 py-2 bg-gray-300 text-gray-800 rounded-md mr-2 itbkk-button-cancel"
+          @click="closeModal"
+        >
+          Cancel
+        </button>
+        <button
+          type="button"
+          class="px-4 py-2 bg-red-500 text-white hover:bg-red-600 rounded-md itbkk-button-confirm"
+          @click="deleteStatus"
+        >
+          Confirm
+        </button>
+      </div>
+    </div>
+  </div>
+  <Toast
+    :show="showToast"
+    :statusCode="statusCode"
+    :operationType="operationType"
+    @close="showToast = false"
+  />
+</template>
 
 <style scoped>
 .modal-mask {
